requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.7
langchain-core==0.3.19

# Telegram
python-telegram-bot==21.6

# Facebook
requests==2.32.3
httpx==0.27.2

# Database
supabase==2.9.1

# Utils
pydantic==2.9.2
pytz==2024.2
